Uint8_t to Uint32_t

+++ include/serial.h
@@ -37,8 +37,8 @@
 #include <stdint.h>
 
 void serial_begin(uint32_t baud);
-void serial_begin_with_pins(uint32_t baud, uint8_t tx, uint8_t rx);
-void serial_begin_with_flow_control(uint32_t baud, uint8_t tx, uint8_t rx, uint8_t rts, uint8_t cts);
+void serial_begin_with_pins(uint32_t baud, uint32_t tx, uint32_t rx);
+void serial_begin_with_flow_control(uint32_t baud, uint32_t tx, uint32_t rx, uint32_t rts, uint32_t cts);
 int serial_available(void);
 size_t serial_println(const char *data);
 size_t serial_write(const char data);
